+++ packages/component-header-footer/src/footer/components/Social/index.test.js
@@ -29,6 +29,7 @@ describe("#Social", () => {
     [`Instagram`, `instagram`],
     [`Youtube`, `youtube`],
     [`LinkedIn`, `linkedin`],
+    [`TikTok`, `tiktok`],
   ];
 
   test.each(cases)(