+++ packages/component-header-footer/src/footer/footer.stories.js
@@ -23,6 +23,7 @@ UnitLogo.args = {
       twitter: "https://twitter.com/?lang=en",
       instagram: "https://instagram.com",
       linkedIn: "https://www.linkedin.com/",
+      tiktok: "https://www.tiktok.com/",
       youtube: "https://www.youtube.com/",
     },
   },
@@ -252,4 +253,10 @@ SixColumns.args = {
       },
     ],
   },
+};
+
+export const Unbranded = Template.bind({});
+Unbranded.args = {
+  ...OneColumn.args,
+  isUnbranded: true,
 };