+++ packages/component-header-footer/types/main.d.ts
@@ -31,12 +31,14 @@ export interface NavTreeProps {
 
 export interface HeaderProps {
   isPartner: boolean;
+  isUnbranded?: boolean;
   navTree: NavTreeProps[];
   title?: string;
   baseUrl?: string;
   parentOrg?:  string;
   parentOrgUrl?: string;
   partnerLogo: Logo;
+  unbrandedLogo?: Logo;
   logo: Logo;
   loggedIn: boolean;
   userName: string;
@@ -75,6 +77,7 @@ export interface Social {
     twitter?: string;
     instagram?: string;
     linkedIn?: string;
+    tiktok?: string;
     youtube?: string;
   };
 }
@@ -87,6 +90,7 @@ export interface Contact {
 }
 
 export interface FooterProps {
+  isUnbranded?: boolean;
   social: Social;
   contact: Contact;
 }