+++ packages/unity-react-core/src/components/ComponentCarousel/core/components/BaseCarousel/index.jsx
@@ -20,14 +20,15 @@ import { calcualteViewItems } from "./helper/width-calculator";
 /** @param {BaseCarouselProps} props */
 const BaseCarousel = ({
   perView = 1,
-  width,
-  maxWidth,
+  width = "",
+  maxWidth = "",
+  // @ts-ignore
   carouselItems,
   cssClass = "",
-  CustomNavComponent,
-  role,
-  ariaLabelledBy,
-  isFullWidth,
+  CustomNavComponent = undefined,
+  role = "",
+  ariaLabelledBy = "",
+  isFullWidth = false,
   removeSideBackground = false,
   hasNavButtons = true,
   hasPositionIndicators = true,
@@ -36,6 +37,7 @@ const BaseCarousel = ({
   hasPeek = true,
   isDraggable = true,
   hasShadow = true,
+  bgColor = "",
 }) => {
   // Only prop for the slider configs we expose is perView. Everything else is
   // considered locked down for Web Standards 2.
@@ -81,7 +83,7 @@ const BaseCarousel = ({
     <div
       role={role}
       aria-labelledby={ariaLabelledBy}
-      className={`glide ${cssClass}`}
+      className={`glide ${cssClass} ${bgColor}`}
       id={instanceName}
       style={{ width, maxWidth }}
       data-remove-side-background={removeSideBackground}
@@ -94,7 +96,9 @@ const BaseCarousel = ({
 
       {CustomNavComponent ? (
         // @ts-ignore
-        <CustomNavComponent instanceName={instanceName} />
+        <CustomNavComponent
+          instanceName={instanceName}
+        />
       ) : (
         <>
           {hasPositionIndicators && <BulletItems buttonCount={buttonCount} />}
@@ -121,6 +125,7 @@ BaseCarousel.propTypes = {
   hasNavButtons: PropTypes.bool,
   hasPositionIndicators: PropTypes.bool,
   imageAutoSize: PropTypes.bool,
+  bgColor: PropTypes.string,
   hasPeek: PropTypes.bool,
   isDraggable: PropTypes.bool,
   hasShadow: PropTypes.bool,