+++ packages/unity-react-core/src/components/Radio/Radio.tsx
@@ -1,259 +1,217 @@
 /* eslint-disable react/jsx-props-no-spreading */
-import React from "react";
+import React, { useState } from "react";
+import { GaEventWrapper } from "../GaEventWrapper/GaEventWrapper";
 
-export const Radios = () => (
-  <form className="uds-form">
-    <fieldset>
-      <legend>A Group of Radios</legend>
-      <div className="form-check">
-        <input
-          className="form-check-input"
-          type="radio"
-          name="exampleRadios"
-          id="exampleRadios1"
-          value="option1"
-          checked
-          data-ga-input="radio button"
-          data-ga-input-name="onclick"
-          data-ga-input-event="select"
-          data-ga-input-action="click"
-          data-ga-input-region="main content"
-          data-ga-input-section="Default radio"
-        />
-        <label className="form-check-label" htmlFor="exampleRadios1">
-          Default radio
-        </label>
-      </div>
-      <div className="form-check">
-        <input
-          className="form-check-input"
-          type="radio"
-          name="exampleRadios"
-          id="exampleRadios2"
-          value="option2"
-          data-ga-input="radio button"
-          data-ga-input-name="onclick"
-          data-ga-input-event="select"
-          data-ga-input-action="click"
-          data-ga-input-region="main content"
-          data-ga-input-section="Second default radio"
-        />
-        <label className="form-check-label" htmlFor="exampleRadios2">
-          Second default radio
-        </label>
-      </div>
-      <div className="form-check">
-        <input
-          className="form-check-input"
-          type="radio"
-          name="exampleRadios"
-          id="exampleRadios3"
-          value="option3"
-          data-ga-input="radio button"
-          data-ga-input-name="onclick"
-          data-ga-input-event="select"
-          data-ga-input-action="click"
-          data-ga-input-region="main content"
-          data-ga-input-section="Third default radio"
-        />
-        <label className="form-check-label" htmlFor="exampleRadios3">
-          Third default radio
-        </label>
-      </div>
-      <div className="form-check disabled">
-        <input
-          className="form-check-input"
-          type="radio"
-          name="exampleRadios"
-          id="exampleRadios4"
-          value="option4"
-          disabled
-          data-ga-input="radio button"
-          data-ga-input-name="onclick"
-          data-ga-input-event="select"
-          data-ga-input-action="click"
-          data-ga-input-region="main content"
-          data-ga-input-section="Disabled radio"
-        />
-        <label className="form-check-label" htmlFor="exampleRadios4">
-          Disabled radio
-        </label>
-      </div>
-    </fieldset>
+interface RadioOption extends React.AriaAttributes {
+  /**
+   * Radio button label
+   */
+  label: string;
+  /**
+   * Input value. Will use the label if not specified.
+   */
+  value?: string;
+  gaEvent?: {
+    // type: string;
+    // ga: string;
+    text: string;
+    name: string;
+    event: string;
+    action: string;
+    region: string;
+    section: string;
+  };
+  /**
+   * Disables a specific radio button
+   */
+  disabled?: boolean;
+}
 
-    <fieldset>
-      <legend>A Group of Invalid Radios</legend>
-      <small id="myInvalidRadiosMsg" className="invalid-feedback is-invalid">
-        <span title="Alert" className="fa fa-icon fa-exclamation-triangle"></span>
-        Form error message
-      </small>
-      <div className="form-check">
-        <input
-          className="form-check-input"
-          type="radio"
-          name="invalidRadios"
-          aria-describedby="myInvalidRadiosMsg"
-          id="invalidRadios1"
-          value="option1"
-          checked
-          data-ga-input="radio button"
-          data-ga-input-name="onclick"
-          data-ga-input-event="select"
-          data-ga-input-action="click"
-          data-ga-input-region="main content"
-          data-ga-input-section="Default radio"
-        />
-        <label className="form-check-label" htmlFor="invalidRadios1">
-          Default radio
-        </label>
-      </div>
-      <div className="form-check">
-        <input
-          className="form-check-input"
-          type="radio"
-          name="invalidRadios"
-          aria-describedby="myInvalidRadiosMsg"
-          id="invalidRadios2"
-          value="option2"
-          data-ga-input="radio button"
-          data-ga-input-name="onclick"
-          data-ga-input-event="select"
-          data-ga-input-action="click"
-          data-ga-input-region="main content"
-          data-ga-input-section="Second default radio"
-        />
-        <label className="form-check-label" htmlFor="invalidRadios2">
-          Second default radio
-        </label>
-      </div>
-      <div className="form-check">
-        <input
-          className="form-check-input"
-          type="radio"
-          name="invalidRadios"
-          aria-describedby="myInvalidRadiosMsg"
-          id="invalidRadios3"
-          value="option3"
-          data-ga-input="radio button"
-          data-ga-input-name="onclick"
-          data-ga-input-event="select"
-          data-ga-input-action="click"
-          data-ga-input-region="main content"
-          data-ga-input-section="Third default radio"
-        />
-        <label className="form-check-label" htmlFor="invalidRadios3">
-          Third default radio
-        </label>
-      </div>
-      <div className="form-check disabled">
-        <input
-          className="form-check-input"
-          type="radio"
-          name="invalidRadios"
-          aria-describedby="myInvalidRadiosMsg"
-          id="invalidRadios4"
-          value="option4"
-          disabled
-          data-ga-input="radio button"
-          data-ga-input-name="onclick"
-          data-ga-input-event="select"
-          data-ga-input-action="click"
-          data-ga-input-region="main content"
-          data-ga-input-section="Disabled radio"
-        />
-        <label className="form-check-label" htmlFor="invalidRadios4">
-          Disabled radio
-        </label>
-      </div>
-    </fieldset>
+export interface RadioProps {
+  /**
+   * Radio button selected state value or default value.
+   * `${id}_option_${index + 1}`
+   */
+  selected?: string;
+  /**
+   * React useState custom setter
+   */
+  setSelected?: React.Dispatch<React.SetStateAction<string>> | undefined;
+  /**
+   * Message for valid input
+   */
+  validMessage?: string;
+  /**
+   * Message for invalid input or errors
+   */
+  invalidMessage?: string;
+  /**
+   * Used for name and id values
+   */
+  id: string;
+  /**
+   * The primary group label
+   */
+  label: string;
+  /**
+   * Disables all radio buttons
+   */
+  disabled?: boolean;
+  /**
+   * Radio button options array
+   */
+  options: Array<RadioOption>;
+}
 
-    <fieldset>
-      <legend>A Group of Valid Radios</legend>
-      <small id="myValidRadiosMsg" className="valid-feedback is-valid">
-        <span
-          title="Success"
-          className="fa fa-icon fa-check-circle">
-        </span>Success
-        message
-      </small>
-      <div className="form-check">
-        <input
-          className="form-check-input"
-          type="radio"
-          name="validRadios"
-          aria-describedby="myValidRadiosMsg"
-          id="validRadios1"
-          value="option1"
-          checked
-          data-ga-input="radio button"
-          data-ga-input-name="onclick"
-          data-ga-input-event="select"
-          data-ga-input-action="click"
-          data-ga-input-region="main content"
-          data-ga-input-section="Default radio"
-        />
-        <label className="form-check-label" htmlFor="validRadios1">
-          Default radio
-        </label>
-      </div>
-      <div className="form-check">
-        <input
-          className="form-check-input"
-          type="radio"
-          name="validRadios"
-          aria-describedby="myValidRadiosMsg"
-          id="validRadios2"
-          value="option2"
-          data-ga-input="radio button"
-          data-ga-input-name="onclick"
-          data-ga-input-event="select"
-          data-ga-input-action="click"
-          data-ga-input-region="main content"
-          data-ga-input-section="Second default radio"
-        />
-        <label className="form-check-label" htmlFor="validRadios2">
-          Second default radio
-        </label>
-      </div>
-      <div className="form-check">
-        <input
-          className="form-check-input"
-          type="radio"
-          name="validRadios"
-          aria-describedby="myValidRadiosMsg"
-          id="validRadios3"
-          value="option3"
-          data-ga-input="radio button"
-          data-ga-input-name="onclick"
-          data-ga-input-event="select"
-          data-ga-input-action="click"
-          data-ga-input-region="main content"
-          data-ga-input-section="Third default radio"
-        />
-        <label className="form-check-label" htmlFor="validRadios3">
-          Third default radio
-        </label>
-      </div>
-      <div className="form-check disabled">
-        <input
-          className="form-check-input"
-          type="radio"
-          name="validRadios"
-          aria-describedby="myValidRadiosMsg"
-          id="validRadios4"
-          value="option4"
-          disabled
-          data-ga-input="radio button"
-          data-ga-input-name="onclick"
-          data-ga-input-event="select"
-          data-ga-input-action="click"
-          data-ga-input-region="main content"
-          data-ga-input-section="Disabled radio"
-        />
-        <label className="form-check-label" htmlFor="validRadios4">
-          Disabled radio
-        </label>
-      </div>
+export const propDefaults: RadioProps = {
+  id: "Radios",
+  label: "",
+  selected: "",
+  setSelected: undefined,
+  // required: false, // Does this component need a required boolean option?
+  disabled: false,
+  validMessage: undefined,
+  invalidMessage: undefined,
+  options: [],
+};
+
+export const Radios: React.FC<RadioProps> = ({
+  id,
+  label,
+  validMessage,
+  invalidMessage,
+  options = [],
+  selected,
+  setSelected,
+  disabled,
+} = propDefaults) => {
+  const [defaultSelectedState, defaultSetSelected] = useState(selected);
+
+  const handleSetSelected = (e: string) => {
+    if (setSelected) {
+      setSelected(e); // custom setter prop
+    } else {
+      defaultSetSelected(e); // default set open function
+    }
+  };
+
+  const getSelectedState = () => {
+    if (setSelected) {
+      return selected; // custom state value
+    } else {
+      return defaultSelectedState; // default state value
+    }
+  };
+
+  const handleOnChange = (e: any, optionValue: string) => {
+    handleSetSelected(optionValue);
+  };
+
+  const defaultGaEvent = {
+    text: "radio button",
+    name: "onclick",
+    event: "select",
+    action: "click",
+    region: "main content",
+    section: "Default radio",
+  };
+
+  const invalidMessageID = `${id}InvalidRadiosMsg`;
+  const validMessageID = `${id}ValidRadiosMsg`;
+
+  const radioButtonGroup = (
+    <fieldset
+      style={invalidMessage || validMessage ? {} : { marginBottom: "3rem" }}
+    >
+      <legend
+        style={invalidMessage || validMessage ? {} : { marginBottom: "-1rem" }}
+      >
+        {label}
+      </legend>
+      {invalidMessage && (
+        <small id={invalidMessageID} className="invalid-feedback is-invalid">
+          <span
+            title="Alert"
+            className="fa fa-icon fa-exclamation-triangle"
+          ></span>
+          {invalidMessage}
+        </small>
+      )}
+      {validMessage && !invalidMessage && (
+        <small id={validMessageID} className="valid-feedback is-valid">
+          <span title="Success" className="fa fa-icon fa-check-circle"></span>
+          {validMessage}
+        </small>
+      )}
+
+      {options.map((val, index) => {
+        const { value, ...ariaProps } = val;
+
+        const idValue = `${id}_option_${index + 1}`;
+
+        // Removes unwanted values from the ariaProps object
+        // ts-ignore was used because some of the type values are not set as optional in the CheckboxOption type definition but
+        // those values should not appear in the ariaProps object and setting required values to optional is not recommended.
+        // Using ...rest instead failed to include the needed aria props and included extra values that aren't needed
+        //@ts-ignore
+        delete ariaProps["label"];
+        delete ariaProps["disabled"];
+        delete ariaProps["gaEvent"];
+        //@ts-ignore
+        delete ariaProps["value"];
+
+        if (!ariaProps["aria-describedby"]) {
+          if (invalidMessage) {
+            ariaProps["aria-describedby"] = invalidMessageID;
+          }
+          if (validMessage && !invalidMessage) {
+            ariaProps["aria-describedby"] = validMessageID;
+          }
+        }
+
+        // Note: label and val.label are different values
+        // same with disabled and val.disabled
+        // one is for the entire group and the
+        // other is for individual radio button options
+        return (
+          <div className="form-radio" key={idValue}>
+            <label htmlFor={idValue}>
+              <GaEventWrapper gaData={defaultGaEvent} prefix="input">
+                <input
+                  tabIndex={0}
+                  type="radio"
+                  name={id}
+                  id={idValue}
+                  value={value ?? val.label}
+                  onChange={e => handleOnChange(e, idValue)}
+                  checked={getSelectedState() === idValue}
+                  disabled={disabled ?? val.disabled ?? false}
+                  {...ariaProps}
+                />
+              </GaEventWrapper>
+              <span>{val.label}</span>
+            </label>
+          </div>
+        );
+      })}
     </fieldset>
-  </form>
-);
+  );
+
+  return radioButtonGroup;
+};
+
+
+// RadioArray and MultiRadioTest can be used to test displaying multiple radio examples on one page
+type RadioArray = {
+  radiosList: Array<RadioProps>;
+};
+
+export const MultiRadioTest: React.FC<RadioArray> = props => {
+  const radioButtonGroup = props.radiosList.map(
+    (val: RadioProps, index: number) => {
+      return <Radios {...val} key={`radioButtonGroup${index}`} />;
+    }
+  );
+
+  return radioButtonGroup;
+};