+++ packages/unity-react-core/src/core/types/shared-types.js
@@ -24,6 +24,7 @@
  * @typedef {Object} ButtonIconOnlyProps
  * @property {Array.<string>} icon
  * @property {string} [color]
+ * @property {boolean} [autoFocus]
  * @property {React.RefObject} [innerRef]
  * @property {function():void} [onClick]
  * @property {"large"|"small"} [size]